Scope Treatments
Added scope treatments to Sintax Tree.
It works by changing the scope according to the function the variable/parameter is in, either a declaration or an activation.
Since this subset does not suport function declarations inside another function, theres no reason to keep apending names to the end of the scope.
Also, scope changes due to control/loop statements are not dealt with.